pyproject.toml
[project]
name = "arcade-next"
version = "0.1.0"
description = "AI-powered software architecture recovery, change, and decay evaluator"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.42.0",
    "tree-sitter>=0.24.0",
    "tree-sitter-java>=0.23.0",
    "networkx>=3.4",
    "jinja2>=3.1",
    "gitpython>=3.1",
]

[project.scripts]
arcade-next = "arcade_next.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
